Fixing the code
In this commit I've changed the 3 parts in which previous developer gave us a puzzle. But with my knowledge of JAVA I could've easily noticed some of the structures. For example: ohmadxhCldo -> replaceChar(because here we've used a method which takes two variables as an input, only one variable replaceChar(.,.) takes as input two variables). After this initial guess everything remaining was just an easy guess(capital letters didn't change while encrypting or decrypting). :)